Bug Fixes and Maven Update and Clean-up - Fix for Server Shops not being breakable (changed where TradeListener was cancelling events) - Moved all Listeners from EventListener to ProtectionListener - Removed a bunch of random debug code - Updated Maven Shade version to 3.1.0 - Updated bstats Repo to codemc - Updated spigot and bukkit to  1.14.4R0.1 - Updated bstats dependency to 1.5
