+++ src/main/java/org/shanerx/tradeshop/TradeShop.java
@@ -34,13 +34,7 @@ import org.shanerx.tradeshop.enumys.Message;
 import org.shanerx.tradeshop.enumys.Setting;
 import org.shanerx.tradeshop.enumys.ShopSign;
 import org.shanerx.tradeshop.framework.AddonManager;
-import org.shanerx.tradeshop.listeners.AddonListener;
-import org.shanerx.tradeshop.listeners.AdminEventListener;
-import org.shanerx.tradeshop.listeners.CustomInventoryListener;
-import org.shanerx.tradeshop.listeners.JoinEventListener;
-import org.shanerx.tradeshop.listeners.ShopCreateListener;
-import org.shanerx.tradeshop.listeners.ShopProtectionListener;
-import org.shanerx.tradeshop.listeners.ShopTradeListener;
+import org.shanerx.tradeshop.listeners.*;
 import org.shanerx.tradeshop.objects.ListManager;
 import org.shanerx.tradeshop.utils.BukkitVersion;
 import org.shanerx.tradeshop.utils.Updater;
@@ -96,7 +90,6 @@ public class TradeShop extends JavaPlugin {
 		lists = new ListManager();
 
 		PluginManager pm = getServer().getPluginManager();
-		pm.registerEvents(new AdminEventListener(this), this);
 		pm.registerEvents(new JoinEventListener(this), this);
 		pm.registerEvents(new ShopProtectionListener(this), this);
 		pm.registerEvents(new ShopCreateListener(), this);
