+++ src/main/java/org/shanerx/tradeshop/listeners/ShopProtectionListener.java
@@ -28,16 +28,24 @@ package org.shanerx.tradeshop.listeners;
 import org.bukkit.Nameable;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
+import org.bukkit.block.BlockState;
 import org.bukkit.block.Sign;
 import org.bukkit.block.data.BlockData;
 import org.bukkit.block.data.Directional;
+import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.EventPriority;
 import org.bukkit.event.Listener;
+import org.bukkit.event.block.Action;
+import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.event.block.BlockPlaceEvent;
 import org.bukkit.event.entity.EntityExplodeEvent;
 import org.bukkit.event.inventory.InventoryMoveItemEvent;
 import org.bukkit.event.inventory.InventoryType;
+import org.bukkit.event.player.PlayerInteractEvent;
 import org.shanerx.tradeshop.TradeShop;
+import org.shanerx.tradeshop.enumys.Message;
+import org.shanerx.tradeshop.enumys.Permissions;
 import org.shanerx.tradeshop.enumys.Setting;
 import org.shanerx.tradeshop.enumys.ShopType;
 import org.shanerx.tradeshop.objects.Shop;
@@ -121,5 +129,142 @@ public class ShopProtectionListener extends Utils implements Listener {
 
 		event.blockList().removeAll(toRemove);
 	}
+
+    @EventHandler(priority = EventPriority.HIGHEST)
+    public void onBlockBreak(BlockBreakEvent event) {
+        Player player = event.getPlayer();
+        Block block = event.getBlock();
+
+        if (ShopType.isShop(block)) {
+            Shop shop = Shop.loadShop((Sign) block.getState());
+            if (shop == null)
+                return;
+
+            if (player.hasPermission(Permissions.ADMIN.getPerm()) || player.getUniqueId().equals(shop.getOwner().getUUID())) {
+
+                if (shop.getStorage() != null) {
+                    new ShopChest(shop.getStorage().getLocation()).resetName();
+                }
+
+                shop.remove();
+                return;
+            }
+            event.setCancelled(true);
+            player.sendMessage(Message.NO_TS_DESTROY.getPrefixed());
+
+        } else if (plugin.getListManager().isInventory(block.getType())) {
+            BlockState bs = block.getState();
+            if (!(bs instanceof Nameable && ((Nameable) bs).getCustomName() != null
+                    && ((Nameable) bs).getCustomName().contains("$ ^Sign:l_"))) {
+                return;
+            }
+
+            if (player.hasPermission(Permissions.ADMIN.getPerm())) {
+                new ShopChest(block.getLocation()).resetName();
+                return;
+            }
+
+            Sign s = findShopSign(block);
+            if (s == null) {
+                new ShopChest(block.getLocation()).resetName();
+                return;
+            }
+
+            if (!ShopType.isShop(s)) {
+                new ShopChest(block.getLocation()).resetName();
+                return;
+            }
+
+            Shop shop = Shop.loadShop(s);
+
+            if (event.getPlayer().getUniqueId().equals(shop.getOwner().getUUID())) {
+                if (!ShopChest.isDoubleChest(block)) {
+                    new ShopChest(shop.getInventoryLocation()).resetName();
+                    shop.removeStorage();
+                    shop.setClosed();
+                    shop.saveShop();
+                } else {
+                    if (bs instanceof Nameable && ((Nameable) bs).getCustomName() != null
+                            && ((Nameable) bs).getCustomName().contains("$ ^Sign:l_")) {
+                        ((Nameable) bs).setCustomName(((Nameable) bs).getCustomName().split("\\$ \\^")[0]);
+
+                        bs.update();
+                    }
+                }
+                return;
+            }
+
+            event.setCancelled(true);
+            player.sendMessage(Message.NO_TS_DESTROY.getPrefixed());
+        }
+    }
+
+    @EventHandler(priority = EventPriority.HIGH)
+    public void onChestOpen(PlayerInteractEvent e) {
+        Block block = e.getClickedBlock();
+
+        if (e.getAction() != Action.RIGHT_CLICK_BLOCK) {
+            return;
+
+        } else if (!plugin.getListManager().isInventory(block.getType())) {
+            return;
+        }
+
+
+        Sign s = findShopSign(block);
+        if (s == null) {
+            return;
+        }
+
+        if (!e.getPlayer().hasPermission(Permissions.ADMIN.getPerm()) && ShopType.isShop(s.getBlock())) {
+            Shop shop = Shop.loadShop(s);
+            if (!shop.getUsersUUID().contains(e.getPlayer().getUniqueId())) {
+                e.getPlayer().sendMessage(Message.NO_TS_OPEN.getPrefixed());
+                e.setCancelled(true);
+            }
+        }
+    }
+
+    @EventHandler(priority = EventPriority.HIGH)
+    public void onBlockPlace(BlockPlaceEvent event) {
+        Block block = event.getBlock();
+
+        if (!plugin.getListManager().getInventories().contains(block.getType()))
+            return;
+
+        if (!ShopChest.isDoubleChest(block)) {
+            Sign shopSign = findShopSign(block);
+
+            if (shopSign == null)
+                return;
+
+            Shop shop = Shop.loadShop(shopSign);
+
+            if (shop.getShopType().isITrade())
+                return;
+
+            new ShopChest(block, shop.getOwner().getUUID(), shopSign.getLocation()).setEventName(event);
+            shop.setInventoryLocation(block.getLocation());
+            shop.saveShop();
+
+        } else {
+            Block otherhalf = ShopChest.getOtherHalfOfDoubleChest(block);
+            Player p = event.getPlayer();
+
+            if (!ShopChest.isShopChest(otherhalf)) {
+                return;
+            }
+
+            ShopChest shopOtherHalf = new ShopChest(otherhalf.getLocation());
+
+            if (shopOtherHalf.hasOwner() && !shopOtherHalf.getOwner().equals(p.getUniqueId())) {
+                event.setCancelled(true);
+                return;
+            }
+
+            new ShopChest(block, shopOtherHalf.getOwner(), shopOtherHalf.getShopSign().getLocation()).setEventName(event);
+        }
+        return;
+    }
 }
